layout/inspector/tests/test_bug1046140.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1046140
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1046140</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  var utils = SpecialPowers.Cc["@mozilla.org/inspector/dom-utils;1"]
    .getService(SpecialPowers.Ci.inIDOMUtils);

  try {
    utils.getSubpropertiesForCSSProperty("--foo");
    ok(false, "expected an exception");
  } catch(e) {
    ok(true, "getSubpropertiesForCSSProperty throws when passed a CSS variable");
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1046140">Mozilla Bug 1046140</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
